Manual merge of hateoas branch, with 'session' concept
OpenstackSession replaces OpenstackClient.  OpenstackSession is lighweight, the Jersey client is heavyweight.
Authentication is stored in the session, and filters are added to the filter chain on every request to bind
in the correct information.  It would be nice to have this information in the properties, but the HEAD method
seems to have a bug (?) in Jersey where these properties aren't set.  Until that can be fixed, we set the filters
on each request.