quantum: Ignore missing properties on deserialization
Some API extensions add properties to the entities. When Jackson sees
these properties it tries to set them on the entity, but since they are
not modelled it fails.

Added a Jackson annotation to ignore such cases, thus API extensions
that aren't modelled yet will not cause an exception in the client.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>