Update ImagesResource.java
X-Image-Meta-Location header has been added to support location option while creating an image.The location option is important to note. It does not copy the entire image into Glance, but reference an original location to where the image can be found. Upon launching an instance of that image, Glance accesses the image from the location specified.