Fix FlavorsResource.showMetadata() 404 Error
Problem: FlavorsResource.showMetadata() results in a 404 error because
it is issuing an API call that does not exist since Flavors do not have
metadata.  Instead, they have “extra-specs”, key-value pairs that the
Nova scheduler uses to launch Servers on Hosts that can supply the
capabilities described by the flavor's extra-specs.

Solution: Replace showMetaData() with showExtraSpecs(), use the correct
“os-extra_specs” extended API path and add a new ExtraSpecs class to
deserialize the API response to a Map<String, String>.