RESTEasyConnector: Isolate the creation of ClientExecutor
The creation of the ClientExecutor is isolated in a protected method,
to allow subclasses to overwrite.

This enables the usage of a customized executor for the request.
The customized executor is needed to set parameters, like timeouts,
on the creation of the executor.

(cherry picked from commit 23922741ab80dc665f83362fbc7b14e8af0240a3)