Let all request path start with "/"
The convention that the path of an OpenStackRequest starts with an
"/" or is empty is introduced. The majority of the code was
written before this change according this convention anyway.
This way the resteasy-connector does not need to add the "/"
between the endpoint and the path. This way the generated URLs
does not contain unintended double "/" anymore, which may
confuse the OpenStack server.